requirements.txt
django
djongo
djangorestframework
graphene-django